requirements.txt
geopandas==1.0.1
openai==1.51.2
pandas==2.2.3
Pillow==10.4.0
python-dotenv==1.0.1
Requests==2.32.3
Shapely==2.0.6
streamlit
